api: - remove historyv1 endpoint in favour of historyv2 endpoint on project level
