+++ api/src/project_view_history.ts
@@ -1,26 +1,123 @@
-import { RequestGenericInterface } from "fastify";
+import { FastifyReply, FastifyRequest, RequestGenericInterface } from "fastify";
 import { AugmentedFastifyInstance } from "types";
-import { VError } from "verror";
 import { AuthenticatedRequest } from "./httpd/lib";
 import { toHttpError } from "./http_errors";
 import * as NotAuthenticated from "./http_errors/not_authenticated";
 import { Ctx } from "./lib/ctx";
+import { safeStringSchema } from "./lib/joiValidation";
 import { isNonemptyString } from "./lib/validation";
 import * as Result from "./result";
-import { BusinessEvent } from "./service/domain/business_event";
+import { businessEventSchema } from "./service/domain/business_event";
 import { ServiceUser } from "./service/domain/organization/service_user";
+import * as History from "./service/domain/workflow/historyFilter";
 import * as Project from "./service/domain/workflow/project";
-import * as Subproject from "./service/domain/workflow/subproject";
+import { ProjectTraceEvent } from "./service/domain/workflow/project_trace_event";
+import Joi = require("joi");
+import VError = require("verror");
+
+const requestBodySchema = Joi.array().items({
+  entityId: Joi.string().required(),
+  entityType: Joi.valid("project").required(),
+  businessEvent: businessEventSchema.required(),
+  snapshot: Joi.object({
+    displayName: safeStringSchema.required(),
+  }).required(),
+});
+
+function validateRequestBody(body): Result.Type<ProjectTraceEvent[]> {
+  const { error, value } = Joi.validate(body, requestBodySchema);
+  return !error ? value : error;
+}
+
+/**
+ * If no filter option is provided the return value is undefined
+ */
+const createFilter = (reply: FastifyReply, request: FastifyRequest): History.Filter | undefined => {
+  const { publisher, startAt, endAt, eventType } = request.query as Querystring;
+  const noFilterSet = !publisher && !startAt && !endAt && !eventType;
+
+  if (noFilterSet) return;
+
+  if (publisher !== undefined) {
+    if (!isNonemptyString(publisher)) {
+      const message = "if present, the query parameter `publisher` must be non-empty string";
+
+      reply.status(400).send({
+        apiVersion: "1.0",
+        error: {
+          code: 400,
+          message,
+        },
+      });
+      request.log.error({ err: message }, "Invalid request body");
+    }
+  }
+
+  // ISO Timestamp example: 01.01.2020 or 2019-12-31T23:00:00.000Z
+  if (startAt !== undefined) {
+    const startAtDate = new Date(startAt);
+    if (isNaN(startAtDate.getTime())) {
+      const message = "if present, the query parameter `startAt` must be a valid ISO timestamp";
+
+      reply.status(400).send({
+        apiVersion: "1.0",
+        error: {
+          code: 400,
+          message,
+        },
+      });
+      request.log.error({ err: message }, "Invalid request body");
+    }
+  }
+
+  if (endAt !== undefined) {
+    const endAtDate = new Date(endAt);
+    if (isNaN(endAtDate.getTime())) {
+      const message = "if present, the query parameter `endAt` must be a valid ISO timestamp";
+
+      reply.status(400).send({
+        apiVersion: "1.0",
+        error: {
+          code: 400,
+          message,
+        },
+      });
+      request.log.error({ err: message }, "Invalid request body");
+    }
+  }
+
+  if (eventType !== undefined) {
+    if (!isNonemptyString(eventType)) {
+      const message = "if present, the query parameter `eventType` must be non-empty string";
+
+      reply.status(400).send({
+        apiVersion: "1.0",
+        error: {
+          code: 400,
+          message,
+        },
+      });
+      request.log.error({ err: message }, "Invalid request body");
+    }
+  }
+
+  return {
+    publisher,
+    startAt,
+    endAt,
+    eventType,
+    // Make typescript happy - noFilterSet condition exists
+  } as History.Filter;
+};
 
 function mkSwaggerSchema(server: AugmentedFastifyInstance) {
   return {
     preValidation: [server.authenticate],
     schema: {
-      deprecated: true,
       description:
         "View the history of a given project (filtered by what the user is allowed to see).",
       tags: ["project"],
-      summary: "View history",
+      summary: "View project history",
       querystring: {
         type: "object",
         properties: {
@@ -39,22 +136,44 @@ function mkSwaggerSchema(server: AugmentedFastifyInstance) {
               "be negative. For example, an `offset` of `-10` with limit `10` requests " +
               "the 10 most recent events.",
           },
+          publisher: {
+            type: "string",
+            description: "Select history entries by the publisher of a new entry",
+          },
+          startAt: {
+            type: "string",
+            description:
+              "Select history entries by date. All entries after this date are shown." +
+              "This is an ISO timestamp",
+          },
+          endAt: {
+            type: "string",
+            description:
+              "Select history entries by date. All entries before this date are shown." +
+              "This is an ISO timestamp",
+          },
+          eventType: {
+            type: "string",
+            description: "Select the history entries by eventType",
+          },
         },
       },
-      security: [
-        {
-          bearerToken: [],
-        },
-      ],
+      security: [{ bearerToken: [] }],
       response: {
         200: {
-          description: "successful response",
+          description: "changes related to the given project in chronological order",
           type: "object",
           properties: {
             apiVersion: { type: "string", example: "1.0" },
             data: {
               type: "object",
               properties: {
+                historyItemsCount: {
+                  type: "number",
+                  description:
+                    "Total number of history items (greater or equal to the number of returned items)",
+                  example: 10,
+                },
                 events: {
                   type: "array",
                   items: {
@@ -82,15 +201,11 @@ function mkSwaggerSchema(server: AugmentedFastifyInstance) {
                         type: "object",
                         additionalProperties: true,
                         properties: {
-                          displayName: { type: "string", example: "townproject" },
+                          displayName: { type: "string", example: "Build a country" },
                         },
                       },
                     },
                   },
-                },
-                historyItemsCount: {
-                  type: "number",
-                  example: 10,
                 },
               },
             },
@@ -102,34 +217,23 @@ function mkSwaggerSchema(server: AugmentedFastifyInstance) {
   };
 }
 
-interface ExposedEvent {
-  entityId: string;
-  entityType: "project" | "subproject";
-  businessEvent: BusinessEvent;
-  snapshot: {
-    displayName?: string;
-  };
-}
-
 interface Service {
-  getProject(ctx: Ctx, user: ServiceUser, projectId: string): Promise<Result.Type<Project.Project>>;
-  getSubprojects(
+  getProjectHistory(
     ctx: Ctx,
     user: ServiceUser,
-    projectId: string,
-  ): Promise<Result.Type<Subproject.Subproject[]>>;
+    projectId: Project.Id,
+    filter?: History.Filter,
+  ): Promise<Result.Type<ProjectTraceEvent[]>>;
 }
 
-interface Request extends RequestGenericInterface {
-  Querystring: {
-    projectId: string;
-    offset?: string;
-    limit?: string;
-    startAt?: string;
-    endAt?: string;
-    publisher?: string;
-    eventType?: string;
-  };
+interface Querystring extends RequestGenericInterface {
+  projectId: string;
+  offset?: string;
+  limit?: string;
+  startAt?: string;
+  endAt?: string;
+  publisher?: string;
+  eventType?: string;
 }
 
 export function addHttpHandler(
@@ -137,7 +241,7 @@ export function addHttpHandler(
   urlPrefix: string,
   service: Service,
 ) {
-  server.get<Request>(
+  server.get<{ Querystring: Querystring }>(
     `${urlPrefix}/project.viewHistory`,
     mkSwaggerSchema(server),
     async (request, reply) => {
@@ -169,10 +273,9 @@ export function addHttpHandler(
       // Default: last created history event
       let offset: number = 0;
       if (request.query.offset !== undefined) {
+        const message = "if present, the query parameter `offset` must be an integer";
         offset = parseInt(request.query.offset, 10);
         if (isNaN(offset)) {
-          const message = "if present, the query parameter `offset` must be an integer";
-
           reply.status(400).send({
             apiVersion: "1.0",
             error: {
@@ -192,7 +295,6 @@ export function addHttpHandler(
         limit = parseInt(request.query.limit, 10);
         if (isNaN(limit) || limit <= 0) {
           const message = "if present, the query parameter `limit` must be a positive integer";
-
           reply.status(400).send({
             apiVersion: "1.0",
             error: {
@@ -206,25 +308,20 @@ export function addHttpHandler(
         }
       }
 
+      const filter = createFilter(reply, request);
+
       try {
-        const projectResult = await service.getProject(ctx, user, projectId);
-        if (Result.isErr(projectResult)) {
-          throw new VError(projectResult, "project.viewHistory failed");
-        }
-        const project: Project.Project = projectResult;
-
-        // Add subprojects' logs to the project log and sort by creation time:
-        const subprojectsResult = await service.getSubprojects(ctx, user, projectId);
-        if (Result.isErr(subprojectsResult)) {
-          throw new VError(subprojectsResult, "project.viewHistory failed");
-        }
-        const subprojects: Subproject.Subproject[] = subprojectsResult;
-        const events: ExposedEvent[] = project.log;
-        for (const subproject of subprojects) {
-          events.push(...subproject.log);
+        // Get all Events in project stream
+        const eventsResult = await service.getProjectHistory(ctx, user, projectId, filter);
+        if (Result.isErr(eventsResult)) {
+          throw new VError(eventsResult, "project.viewHistory failed");
         }
 
-        events.sort(byEventTime);
+        const eventsResultVerified = validateRequestBody(eventsResult);
+        if (Result.isErr(eventsResultVerified)) {
+          throw new VError(eventsResultVerified, "project.viewHistory failed");
+        }
+        const events: ProjectTraceEvent[] = eventsResultVerified;
 
         const offsetIndex = offset < 0 ? Math.max(0, events.length + offset) : offset;
         const slice = events.slice(
@@ -236,8 +333,8 @@ export function addHttpHandler(
         const body = {
           apiVersion: "1.0",
           data: {
-            events: slice,
             historyItemsCount: events.length,
+            events: slice,
           },
         };
         reply.status(code).send(body);
@@ -248,12 +345,4 @@ export function addHttpHandler(
       }
     },
   );
-}
-
-function byEventTime(a: ExposedEvent, b: ExposedEvent): -1 | 0 | 1 {
-  const timeA = new Date(a.businessEvent.time);
-  const timeB = new Date(b.businessEvent.time);
-  if (timeA < timeB) return -1;
-  if (timeA > timeB) return 1;
-  return 0;
 }
